Fix ESLint in UI
I've renamed the eslintrc file to use `.cjs` as this is the easiest way to get it running in a `"type": "module"` setup. I can't recall if there are better ways, but I just want linting working for now.

I've also switched the custom eslint package it uses to `library`.