add `Box` and a large chunk of theme configuration
Much of this theme configuration is for the `Box` component and upcoming components. There is a lot to start with, and I wanted it to be chunked up a little more but there is a lot of reliance on theme utilities and configuration files that is difficult to break apart without some work.

There was also a lot of formatting changes from my editor because I have not set up Prettier yet, and I needed some standardization going.