add `Card` and `Stack` components, with tests and stories.
This also removes the old `Card` component that came with the starter, as well as adding some required packages for `Stack` and a utility function: `flattenFragments`. The `react-is` package is required to accomplish this with any ease, because fragments are actually quite complicated to render based on props without getting funky behavior with empty child wrappers and other issues.

I learned all this the hard way, and it was pain.