remove unused lint disable line in web app
I don't know what is up with this rule. My linter constantly says it is not needed, but when I removed it before I would get a lint error for running an unsafe function.

I imagine it has something to do with side effects, but it isn't complaining now, so I'm getting rid of the ignore rule.