add `BlockQuote` component
I ran into a little snag on this one, apparently Next.js server components are not compatible with Stitches `css` styles because you have to run a function in a component, which only server components are allowed to do.

I just converted it to use `styled` but it was a bit annoying.